+++ src/components/navigation/NavBar.js
@@ -22,7 +22,13 @@ export default function NavBar() {
   }
   return (
     <>
-      <Navbar collapseOnSelect expand='lg' bg='dark' variant='dark'>
+      <Navbar
+        collapseOnSelect
+        expand='lg'
+        bg='dark'
+        variant='dark'
+        className='fixed-top'
+      >
         <Container>
           <div>
             <span className='title'>